requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy == 1.23.4
pillow == 9.3.0
torch == 1.13.0
torchvision == 0.14.0

wandb~=0.13.5